+++ test/unit/wenet/model/callback_message/test_event.py
@@ -2,13 +2,7 @@ from __future__ import absolute_import, annotations
 
 from unittest import TestCase
 
-from wenet.model.callback_message.event import WeNetAuthenticationEvent, Event
-
-
-class TestWeNetAuthenticationEvent(TestCase):
-    def test_repr(self):
-        auth_event = WeNetAuthenticationEvent("external id", "code")
-        self.assertEqual(auth_event, WeNetAuthenticationEvent.from_repr(auth_event.to_repr()))
+from wenet.model.callback_message.event import Event
 
 
 class TestEvent(TestCase):